fix small int warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@4238 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainApps/Enviro/wx/LocationDlg.cpp
@@ -185,7 +185,7 @@ void LocationDlg::UpdateSlider()
 		// time slider
 		float fTotalTime = GetAnim(m_iAnim)->GetLastTime();
 		float fTime = GetEngine(m_iAnim)->GetTime();
-		m_iPos = (fTime / fTotalTime * 1000);
+		m_iPos = (int) (fTime / fTotalTime * 1000);
 		GetAnimPos()->SetValue(m_iPos);
 	}
 }
